Create challenge5.sol
To perform a stress test of a for loop in Solidity, we can create two different functions:

A function that performs a simple operation  (like adding numbers) in 1000 loops.
A function that performs a complex operation (like calculating product of numbers) in 100 loops.
The  'simple' function executes successfully without errors, which indicates that performing simple operations in 1000 loops is within the gas limit(300k) whereas the 'complex' function throws an error, which is likely due to exceeding the gas limit(300k). In Solidity, running out of gas results in a transaction failure. This is particularly common in loops with high iteration counts or complex computations.